fix: refactor a Mapper de Espacio y adaptacion a la nueva DB de Espacio y EspacioOcupación

+++ src/main/java/com/coagronet/espacio/repositories/EspacioRepository.java
@@ -7,12 +7,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.coagronet.espacio.Espacio;
 
-public interface EspacioRepository extends JpaRepository<Espacio, Integer> {
+public interface EspacioRepository extends JpaRepository<Espacio, Long> {
+
 
     List<Espacio> findByBloqueSedeEmpresaIdAndBloqueIdAndEstadoIdNotOrderByIdAsc(Long empresaId, Integer bloqueId,
             Integer estadoId);
 
-    Optional<Espacio> findByIdAndBloqueSedeEmpresaIdAndEstadoIdNot(Integer id, Long empresaId, Integer estadoId);
+    Optional<Espacio> findByIdAndBloqueSedeEmpresaIdAndEstadoIdNot(Long id, Long empresaId, Integer estadoId);
 
-    boolean existsByIdAndBloqueSedeEmpresaIdAndEstadoIdNot(Integer id, Long empresaId, Integer estadoId);
+    boolean existsByIdAndBloqueSedeEmpresaIdAndEstadoIdNot(Long id, Long empresaId, Integer estadoId);
 }
